pyproject.toml
[project]
name = "terms"
version = "0.1.0"
description = ""
authors = [
    {name = "Maicol Rodrigues"}
]
readme = "README.md"
requires-python = ">=3.12,<4"
dependencies = [
    "bitsandbytes (>=0.46,<0.47) ; sys_platform == 'linux'",
    "transformers (>=4.53.0,<5.0.0)",
    "pytorch-lightning (>=2.5.2,<3.0.0)",
    "datasets (>=3.6.0,<4.0.0)",
    "pandas (>=2.3.0,<3.0.0)",
    "torch (>=2.7.1,<3.0.0)",
    "peft (>=0.15.2,<0.16.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "pandera (>=0.24.0,<0.25.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "mlflow (>=3.1.1,<4.0.0)",
    "graphql-core (>=3.2.6,<4.0.0)",
    "scikit-learn (>=1.7.0,<2.0.0)",
    "pyarrow (>=20.0.0,<21.0.0)",
    "fastapi (>=0.115.14,<0.116.0)",
    "typer (>=0.16.0,<0.17.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
]

[project.scripts]
terms = "terms.cli:app" 

[tool.poetry]
packages = [{ include = "terms", from = "src" }]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
openpyxl = "^3.1.5"
ipykernel = "^6.29.5"

